+++ build-conventions/src/main/java/org/elasticsearch/gradle/internal/conventions/precommit/FormattingPrecommitPlugin.java
@@ -89,12 +89,15 @@ public class FormattingPrecommitPlugin implements Plugin<Project> {
             return new String[] {
                 "src/*/java/org/elasticsearch/action/admin/cluster/repositories/**/*.java",
                 "src/*/java/org/elasticsearch/action/admin/cluster/snapshots/**/*.java",
-                "src/test/java/org/elasticsearch/common/xcontent/support/AbstractFilteringTestCase.java",
                 "src/test/java/org/elasticsearch/common/xcontent/support/XContentMapValuesTests.java",
                 "src/*/java/org/elasticsearch/index/snapshots/**/*.java",
                 "src/*/java/org/elasticsearch/repositories/**/*.java",
                 "src/*/java/org/elasticsearch/search/aggregations/**/*.java",
                 "src/*/java/org/elasticsearch/snapshots/**/*.java" };
+        } else if (projectPath.equals(":test:framework")) {
+            return new String[] {
+                "src/test/java/org/elasticsearch/common/xcontent/support/AbstractFilteringTestCase.java",
+            };
         } else {
             // Normally this isn"t necessary, but we have Java sources in
             // non-standard places
@@ -200,7 +203,6 @@ public class FormattingPrecommitPlugin implements Plugin<Project> {
         ":test:fixtures:geoip-fixture",
         ":test:fixtures:krb5kdc-fixture",
         ":test:fixtures:old-elasticsearch",
-        ":test:framework",
         ":test:logger-usage",
         ":x-pack:docs",
         ":x-pack:license-tools",
