+++ libs/x-content/src/test/java/org/elasticsearch/xcontent/support/filtering/JsonXContentFilteringTests.java
@@ -6,20 +6,24 @@
  * Side Public License, v 1.
  */
 
-package org.elasticsearch.common.xcontent.support.filtering;
+package org.elasticsearch.xcontent.support.filtering;
 
 import org.elasticsearch.xcontent.XContentBuilder;
 import org.elasticsearch.xcontent.XContentType;
 
-public class SmileFilteringGeneratorTests extends AbstractXContentFilteringTestCase {
+public class JsonXContentFilteringTests extends AbstractXContentFilteringTestCase {
 
     @Override
     protected XContentType getXContentType() {
-        return XContentType.SMILE;
+        return XContentType.JSON;
     }
 
     @Override
     protected void assertFilterResult(XContentBuilder expected, XContentBuilder actual) {
-        assertXContentBuilderAsBytes(expected, actual);
+        if (randomBoolean()) {
+            assertXContentBuilderAsString(expected, actual);
+        } else {
+            assertXContentBuilderAsBytes(expected, actual);
+        }
     }
 }
