+++ test/framework/src/main/resources/org/elasticsearch/common/xcontent/support/many_filters.txt
