+++ test/framework/src/main/resources/org/elasticsearch/common/xcontent/support/sample_just_authors.json
