+++ test/framework/src/main/resources/org/elasticsearch/common/xcontent/support/sample_no_properties_en_names.json
