+++ test/framework/src/main/resources/org/elasticsearch/common/xcontent/support/sample_no_title.json
